Move old Lambda provider implementation to legacy package
