+++ localstack/services/lambda_/event_source_listeners/sqs_event_source_listener.py
@@ -11,11 +11,11 @@ from localstack.services.lambda_.event_source_listeners.adapters import (
 from localstack.services.lambda_.event_source_listeners.event_source_listener import (
     EventSourceListener,
 )
-from localstack.services.lambda_.lambda_executors import InvocationResult
 from localstack.services.lambda_.lambda_utils import (
     filter_stream_records,
     message_attributes_to_lower,
 )
+from localstack.services.lambda_.legacy.lambda_executors import InvocationResult
 from localstack.utils.aws import arns
 from localstack.utils.aws.arns import extract_region_from_arn
 from localstack.utils.threads import FuncThread
