+++ localstack/services/lambda_/legacy/lambda_starter.py
@@ -5,8 +5,8 @@ from moto.awslambda import models as moto_awslambda_models
 from localstack import config
 from localstack.aws.connect import connect_to
 from localstack.services.edge import ROUTER
-from localstack.services.lambda_.lambda_api import handle_lambda_url_invocation
 from localstack.services.lambda_.lambda_utils import get_default_executor_mode
+from localstack.services.lambda_.legacy.lambda_api import handle_lambda_url_invocation
 from localstack.services.plugins import ServiceLifecycleHook
 from localstack.utils.analytics import log
 from localstack.utils.aws import arns
@@ -44,7 +44,8 @@ class LambdaLifecycleHook(ServiceLifecycleHook):
 
 def start_lambda(port=None, asynchronous=False):
     from localstack.services.infra import start_local_api
-    from localstack.services.lambda_ import lambda_api, lambda_utils
+    from localstack.services.lambda_ import lambda_utils
+    from localstack.services.lambda_.legacy import lambda_api
 
     log.event(
         "lambda:config",
@@ -71,7 +72,7 @@ def start_lambda(port=None, asynchronous=False):
 
 
 def stop_lambda() -> None:
-    from localstack.services.lambda_.lambda_api import cleanup
+    from localstack.services.lambda_.legacy.lambda_api import cleanup
 
     """
     Stops / cleans up the Lambda Executor
