+++ localstack/services/lambda_/urlrouter.py
@@ -1,3 +1,4 @@
+"""Routing for Lambda function URLs: https://docs.aws.amazon.com/lambda/latest/dg/lambda-urls.html"""
 import base64
 import json
 import logging
