+++ tests/aws/services/lambda_/test_lambda_integration_dynamodbstreams.py
@@ -5,8 +5,8 @@ import time
 import pytest
 
 from localstack.aws.api.lambda_ import Runtime
-from localstack.services.lambda_.lambda_api import INVALID_PARAMETER_VALUE_EXCEPTION
 from localstack.services.lambda_.lambda_utils import LAMBDA_RUNTIME_PYTHON39
+from localstack.services.lambda_.legacy.lambda_api import INVALID_PARAMETER_VALUE_EXCEPTION
 from localstack.testing.aws.lambda_utils import (
     _await_dynamodb_table_active,
     _await_event_source_mapping_enabled,
