+++ tests/aws/services/lambda_/test_lambda_integration_sqs.py
@@ -6,13 +6,13 @@ import pytest
 from botocore.exceptions import ClientError
 
 from localstack.aws.api.lambda_ import Runtime
-from localstack.services.lambda_.lambda_api import (
-    BATCH_SIZE_RANGES,
-    INVALID_PARAMETER_VALUE_EXCEPTION,
-)
 from localstack.services.lambda_.lambda_utils import (
     LAMBDA_RUNTIME_PYTHON38,
     LAMBDA_RUNTIME_PYTHON39,
+)
+from localstack.services.lambda_.legacy.lambda_api import (
+    BATCH_SIZE_RANGES,
+    INVALID_PARAMETER_VALUE_EXCEPTION,
 )
 from localstack.testing.aws.lambda_utils import _await_event_source_mapping_enabled, is_old_provider
 from localstack.testing.aws.util import is_aws_cloud
